+++ src/stats/export/mod.rs
@@ -26,6 +26,7 @@ mod pressure;
 mod rotation;
 mod rush;
 mod speed_flip;
+mod territorial_pressure;
 mod touch;
 mod wall_aerial;
 mod wall_aerial_shot;